Handling grid better.
The grid will be placed every approximately x_grid_by and y_grid_by pixels.
This is approximate, because the grid will be stretched in order to generate
equally distributed grid lines, and at least two grid lines should be located at
the edges.
For example, if the x_grid_by = 256 and the map_width = 512 the 3 grid lines
will be generated: at x = 1, x=257 and x=512. In these values, the canvas_margin
size is added. If the x_grid_by = 256 and the map_width = 544, then the x_lines will
be placed at 1, 273, 544.